pyproject.toml
[tool.poetry]
name = "simpleaitranslatoruserinterface"
version = "0.1.0"
description = ""
authors = ["Adam Pawełek <adam.pwk@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.37.1"
simpleaitranslator = "^2.1.5"
python-dotenv = "^1.0.1"
htbuilder = "^0.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
